project_leveler/requirements.txt
fastapi
uvicorn
python-dotenv
openai
python-multipart
pypdf
